feat: add TypeScript template property fields

+++ spring-ddd-infrastructure-persistence/src/main/java/com/springddd/infrastructure/persistence/entity/GenColumnsEntity.java
@@ -41,7 +41,7 @@ public class GenColumnsEntity {
 
     private Integer tableFilterType;
 
-    private Integer formComponent;
+    private Byte formComponent;
 
     private Boolean formVisible;
